create roomを修正
